added created language picker and darkmode toggler
