+++ src/components/NavBar/NavBar.scss
